Overflow change, change to new block
Overflow changed to 0, add new code for the when it needs to create a new block. Tried creating a new block then adding everything to in then using the put back function. But that didn't work. So I left it as a copy and past from the normal push. And it seems to be the farthest I can get to writing.